+++ src/main/java/web/dao/UserDao.java
@@ -13,7 +13,7 @@ public interface UserDao {
 
     void saveUser(User user);
 
-    User getUser(int id);
+    User getUser(long id);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
